backend/requirements.txt
Django
djangorestframework
psycopg2
pdfkit
django-cors-headers
